+++ lavrentev.makar/S1/test-List.cpp
@@ -5,39 +5,39 @@
 
 BOOST_AUTO_TEST_CASE(clear_test)
 {
-  lavrentev::List<int> k{};
-  lavrentev::LIter<int> it;
+  lavrentev::List<size_t> k{};
+  lavrentev::LIter<size_t> it;
   it = k.insert(it, 3);
   it = k.insert(it, 4);
   k.clear();
-  BOOST_CHECK(k.cbegin() == lavrentev::LCIter<int>());
+  BOOST_CHECK(k.cbegin() == lavrentev::LCIter<size_t>());
 }
 
 BOOST_AUTO_TEST_CASE(insert_test)
 {
-  lavrentev::List<int> k{};
-  lavrentev::LIter<int> it;
+  lavrentev::List<size_t> k{};
+  lavrentev::LIter<size_t> it;
   it = k.insert(it, 1);
   BOOST_TEST(k.front() == 1);
 }
 
 BOOST_AUTO_TEST_CASE(popFront_test)
 {
-  lavrentev::List<int> k{};
-  BOOST_CHECK(k.popFront() == lavrentev::LIter<int>{});
-  lavrentev::LIter<int> it;
+  lavrentev::List<size_t> k{};
+  BOOST_CHECK(k.popFront() == lavrentev::LIter<size_t>{});
+  lavrentev::LIter<size_t> it;
   it = k.insert(it, 1);
   it = k.insert(it, 2);
   BOOST_CHECK(k.popFront() == it);
-  BOOST_CHECK(k.popFront() == lavrentev::LIter<int>{});
+  BOOST_CHECK(k.popFront() == lavrentev::LIter<size_t>{});
 }
 
 BOOST_AUTO_TEST_CASE(begin_test)
 {
-  lavrentev::List<int> k{};
-  BOOST_CHECK(k.begin() == lavrentev::LIter<int>());
+  lavrentev::List<size_t> k{};
+  BOOST_CHECK(k.begin() == lavrentev::LIter<size_t>());
 
-  lavrentev::LIter<int> it;
+  lavrentev::LIter<size_t> it;
   it = k.insert(it, 3);
   BOOST_CHECK(k.begin() == it);
   BOOST_TEST(*k.begin() == 3);
@@ -45,47 +45,47 @@ BOOST_AUTO_TEST_CASE(begin_test)
 
 BOOST_AUTO_TEST_CASE(cbegin_test)
 {
-  lavrentev::List<int> k{};
-  BOOST_CHECK(k.cbegin() == lavrentev::LCIter<int>());
+  lavrentev::List<size_t> k{};
+  BOOST_CHECK(k.cbegin() == lavrentev::LCIter<size_t>());
 
-  lavrentev::LIter<int> it;
+  lavrentev::LIter<size_t> it;
   it = k.insert(it, 3);
-  lavrentev::LCIter<int> cit = k.cbegin();
-  BOOST_CHECK(cit != lavrentev::LCIter<int>());
+  lavrentev::LCIter<size_t> cit = k.cbegin();
+  BOOST_CHECK(cit != lavrentev::LCIter<size_t>());
   BOOST_TEST(*cit == 3);
 }
 
 BOOST_AUTO_TEST_CASE(end_test)
 {
-  lavrentev::List<int> k{};
-  BOOST_CHECK(k.end() == lavrentev::LIter<int>());
+  lavrentev::List<size_t> k{};
+  BOOST_CHECK(k.end() == lavrentev::LIter<size_t>());
 }
 
 BOOST_AUTO_TEST_CASE(cend_test)
 {
-  lavrentev::List<int> k{};
-  BOOST_CHECK(k.cend() == lavrentev::LCIter<int>());
+  lavrentev::List<size_t> k{};
+  BOOST_CHECK(k.cend() == lavrentev::LCIter<size_t>());
 }
 
 BOOST_AUTO_TEST_CASE(front_test)
 {
-  lavrentev::List<int> k{};
+  lavrentev::List<size_t> k{};
   BOOST_CHECK_THROW(k.front(), std::out_of_range);
 
-  lavrentev::LIter<int> it;
+  lavrentev::LIter<size_t> it;
   it = k.insert(it, 3);
   BOOST_TEST(k.front() == 3);
 
-  const lavrentev::List<int> &ck = k;
+  const lavrentev::List<size_t> &ck = k;
   BOOST_TEST(ck.front() == 3);
 }
 
 BOOST_AUTO_TEST_CASE(empty_test)
 {
-  lavrentev::List<int> k{};
+  lavrentev::List<size_t> k{};
   BOOST_TEST(k.empty());
 
-  lavrentev::LIter<int> it;
+  lavrentev::LIter<size_t> it;
   it = k.insert(it, 3);
   BOOST_TEST(!k.empty());
 }
